requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
pandas
tqdm
gdown
wandb
pillow
scikit-learn

librosa
soundfile
